refactor array problems - segregate array problems into easy, medium and hard
