+++ src/strivers/arrays/easy/LeftRotateArrayByDPlaces.java
@@ -1,4 +1,4 @@
-package strivers.arrays;
+package strivers.arrays.easy;
 
 import util.ArrayUtil;
