+++ src/strivers/arrays/easy/MaxConsecutiveOnes.java
@@ -1,4 +1,4 @@
-package strivers.arrays;
+package strivers.arrays.easy;
 
 public class MaxConsecutiveOnes {
