+++ src/strivers/arrays/easy/MissingNumberInOneToN.java
@@ -1,4 +1,4 @@
-package strivers.arrays;
+package strivers.arrays.easy;
 
 public class MissingNumberInOneToN {
