+++ src/strivers/arrays/easy/MoveZeroesToEnd.java
@@ -1,4 +1,4 @@
-package strivers.arrays;
+package strivers.arrays.easy;
 
 import util.ArrayUtil;
